move degree requirement scraping to catolog.ts and create typedefs
